add basic scoreUpdate event
Whenever a new score is received by the server, it should broadcast the update to everyone who is connected.  So it does that now.

The assumption here is that we will keep the score update having the same structure on broadcast as it does when a client sends a score update to the server.  So then the client will take the score and change what is needed instead of the server, which just logs the update in the db and then reflects it back to everyone.