user script basic functionality
The user script now sends players at the start of the game, and sends
the game time on every score, then displays a simple scoreboard on the
screen.  It now also should function as a user script.  The server side
code was updated to include the team names on broadcast.